ComputeShader: Update arrayTexture demo
